+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 0.6
 - Added `Book::revise_chapter1` which represents a totally new system for interactively revising chapter content, analyzing for bad patterns and then, either altering the text content, removing the match entirely, or simply ignoring the match.
-- Includes a 'batch method' which allows the user to choose either 'alter', 'remove', or 'ignore' for ALL found bad patterns at once and addresses them in a batch instead of interactively.
+- Includes 3 modes:
+  - Mode 1: Apply an operation to ALL instances of bad pattern matches at once.
+  - Mode 2: Iterate through each bad pattern and choose an operation to apply to all of the matches.
+  - Mode 3: Interactively go through each bad pattern match and decide which operation to apply.
 - Added a `Utils` module for extra goodies
 
 ## 0.5 - 2024.11.27